Moved BaseSignTest to the base package.| #1947
