+++ FlowCrypt/src/androidTest/java/com/flowcrypt/email/ui/activity/enterprise/MainSignInFragmentEnterpriseTest.kt
@@ -31,7 +31,7 @@ import com.flowcrypt.email.rules.FlowCryptMockWebServerRule
 import com.flowcrypt.email.rules.RetryRule
 import com.flowcrypt.email.rules.ScreenshotTestRule
 import com.flowcrypt.email.ui.activity.MainActivity
-import com.flowcrypt.email.ui.activity.base.BaseSignTest
+import com.flowcrypt.email.ui.base.BaseSignTest
 import com.flowcrypt.email.util.PrivateKeysManager
 import com.flowcrypt.email.util.TestGeneralUtil
 import com.flowcrypt.email.util.exception.ApiException
